+++ mirrord/agent/src/steal/ip_tables.rs
@@ -1,6 +1,8 @@
 use mirrord_protocol::Port;
 use nix::unistd::getgid;
 use rand::distributions::{Alphanumeric, DistString};
+use tokio::process::Command;
+use tracing::warn;
 
 use crate::error::{AgentError, Result};
 
@@ -62,6 +64,7 @@ pub(crate) struct SafeIpTables<IPT: IPTables> {
     inner: IPT,
     chain_name: String,
     formatter: IPTableFormatter,
+    flush_connections: bool,
 }
 
 const IPTABLES_TABLE_NAME: &str = "nat";
@@ -74,7 +77,7 @@ impl<IPT> SafeIpTables<IPT>
 where
     IPT: IPTables,
 {
-    pub(super) fn new(ipt: IPT) -> Result<Self> {
+    pub(super) fn new(ipt: IPT, flush_connections: bool) -> Result<Self> {
         let formatter = IPTableFormatter::detect(&ipt)?;
 
         let chain_name = Self::get_chain_name();
@@ -87,6 +90,7 @@ where
             inner: ipt,
             chain_name,
             formatter,
+            flush_connections,
         })
     }
 
@@ -167,13 +171,32 @@ where
 
     /// Adds port redirection, and bypass gid packets from iptables.
     #[tracing::instrument(level = "trace", skip(self))]
-    pub(super) fn add_stealer_iptables_rules(
+    pub(super) async fn add_stealer_iptables_rules(
         &self,
         redirected_port: Port,
         target_port: Port,
     ) -> Result<()> {
         self.add_redirect(redirected_port, target_port)
-            .and_then(|_| self.add_bypass_own_packets())
+            .and_then(|_| self.add_bypass_own_packets())?;
+
+        if self.flush_connections {
+            let conntrack = Command::new("conntrack")
+                .args([
+                    "--delete",
+                    "--proto",
+                    "tcp",
+                    "--orig-port-dst",
+                    &target_port.to_string(),
+                ])
+                .output()
+                .await?;
+
+            if !conntrack.status.success() && conntrack.status.code() != Some(256) {
+                warn!("`conntrack` output is {conntrack:#?}");
+            }
+        }
+
+        Ok(())
     }
 
     /// Removes port redirection, and bypass gid packets from iptables.
@@ -301,7 +324,7 @@ mod tests {
             .times(1)
             .returning(|_, _| Ok(()));
 
-        let ipt = SafeIpTables::new(mock).expect("Create Failed");
+        let ipt = SafeIpTables::new(mock, false).expect("Create Failed");
 
         assert!(ipt.add_redirect(69, 420).is_ok());
 
@@ -353,7 +376,7 @@ mod tests {
             .times(1)
             .returning(|_, _| Ok(()));
 
-        let ipt = SafeIpTables::new(mock).expect("Create Failed");
+        let ipt = SafeIpTables::new(mock, false).expect("Create Failed");
 
         assert!(ipt.add_redirect(69, 420).is_ok());