+++ mirrord/layer/src/file/filter.rs
@@ -100,6 +100,8 @@ fn generate_remote_ro_set() -> RegexSet {
         // AWS cli cache
         // "file not exist" for identity caches (AWS)
         r".aws/cli/cache/.+\.json$",
+        r".aws/credentials$",
+        r".aws/config$",
         // for dns resolving
         r"^/etc/resolv.conf$",
         r"^/etc/hosts$",